userlogin finish?

+++ js/register.js
@@ -1,4 +1,4 @@
-let testUsers = [];
+let users = [];
 
 async function initRegister() {
     await loadUsers();
@@ -10,8 +10,14 @@ async function register() {
         email: inputEmail.value,
         password: inputPassword.value,
     };
-    users.push(newUser);
-    await setItem('users', JSON.stringify(users));
+    let userExists = users.find(user => user.email === newUser.email && user.name === newUser.name);
+    if (userExists) {
+        console.log("Benutzer existiert bereits:", userExists.name);
+    } else {
+        console.log("Kein Benutzer mit diesen Daten gefunden.\nNeuer Nutzer wird angelegt: ", newUser.name);
+        users.push(newUser);
+        await setItem('users', JSON.stringify(users));
+    }
     await loadUsers();
 }
 
@@ -19,23 +25,19 @@ function backToLogIn() {
     window.location.href = 'log-in.html';
 }
 
-function guestLogIn() {
-    window.location.href = '../../html/add_task.html';
-}
-
 async function loadUsers() {
     try {
         let loadedUsers = JSON.parse(await getItem('users'));
         if (Array.isArray(loadedUsers)) {
-            testUsers = loadedUsers;
+            users = loadedUsers;
         } else {
             console.log("Keine Benutzer gefunden, Initialisierung mit einem leeren Array.");
-            testUsers = [];
+            users = [];
         }
     } catch (error) {
         console.error("Fehler beim Laden der Benutzer: ", error);
-        testUsers = [];
+        users = [];
     }
 
-    console.log("Alle geladenen Benutzer: ", testUsers);
+    console.log("Alle geladenen Benutzer: ", users);
 }
